Add new tab(setting)
