+++ src/gui/pages/__init__.py
@@ -1,6 +1,7 @@
 __all__ = [
-    'PageKeyboard', 'PageSelectCamera','PageAbout'
+    'PageKeyboard', 'PageSelectCamera','PageAbout', 'PageSetting'
 ]
 from .page_keyboard import PageKeyboard
 from .page_select_camera import PageSelectCamera
 from .page_about import PageAbout
+from .page_setting import PageSetting
